+++ python/dalex/dalex/_explainer/object.py
@@ -188,7 +188,8 @@ class Explainer:
                         variables=None,
                         grid_points=101,
                         variable_splits=None,
-                        processes=1):
+                        processes=1,
+                        disable=False):
 
         """Creates CeterisParibus object
 
@@ -199,6 +200,7 @@ class Explainer:
         :param grid_points: number of points in a single variable split if calculated automatically
         :param variable_splits: mapping of variables into points the profile will be calculated, if None then calculate with the function `_calculate_variable_splits`
         :param processes: integer, number of parallel processes, iterated over variables
+        :param disable: disable tqdm progress bar printing
         :return CeterisParibus object
         """
 
@@ -213,7 +215,7 @@ class Explainer:
                 processes=processes
             )
 
-        predict_profile_.fit(self, new_observation, y)
+        predict_profile_.fit(self, new_observation, y, disable)
 
         return predict_profile_
 
@@ -242,7 +244,7 @@ class Explainer:
         return model_performance_
 
     def model_parts(self,
-                    loss_function='rmse',
+                    loss_function=None,
                     type=('variable_importance', 'ratio', 'difference'),
                     N=None,
                     B=10,
@@ -254,7 +256,7 @@ class Explainer:
 
         """Creates VariableImportance object
 
-        :param loss_function: a function that will be used to assess variable importance
+        :param loss_function: str or a function that will be used to assess variable importance, e.g. 'rmse' or '1-auc'
         :param type: 'variable_importance'/'ratio'/'difference' type of transformation that should be applied for dropout loss
         :param N: number of observations that should be sampled for calculation of variable importance
         :param B: number of permutation rounds to perform on each variable
@@ -268,6 +270,8 @@ class Explainer:
 
         types = ('variable_importance', 'ratio', 'difference')
         type = check_method_type(type, types)
+
+        loss_function = check_loss_function(self, loss_function)
 
         model_parts_ = VariableImportance(
             loss_function=loss_function,
@@ -295,7 +299,8 @@ class Explainer:
                       grid_points=101,
                       intercept=True,
                       processes=1,
-                      random_state=None):
+                      random_state=None,
+                      disable=False):
 
         """Dataset Level Variable Effect as Partial Dependency Profile or Accumulated Local Effects
 
@@ -310,6 +315,7 @@ class Explainer:
         :param intercept: False if center data on 0
         :param processes: integer, number of parallel processes, iterated over variables
         :param random_state: int, seed for random number generator
+        :param disable: disable tqdm progress bar printing
         :return: VariableEffect object
         """
 
@@ -327,7 +333,7 @@ class Explainer:
         I = np.random.choice(np.arange(N), N, replace=False)
 
         ceteris_paribus = CeterisParibus(grid_points=grid_points, processes=processes)
-        ceteris_paribus.fit(self, self.data.iloc[I, :], self.y[I])
+        ceteris_paribus.fit(self, self.data.iloc[I, :], self.y[I], disable=disable)
 
         model_profile_ = AggregatedProfiles(
             type=type,
@@ -339,7 +345,7 @@ class Explainer:
             random_state=random_state
         )
 
-        model_profile_.fit(ceteris_paribus)
+        model_profile_.fit(ceteris_paribus, disable)
 
         return model_profile_